requirements.txt
crewai>=0.28.8
langchain-openai>=0.1.7
python-decouple>=3.8
openai>=1.30.0